src/main/java/br/edu/ifal/leilaogqso/Avaliador.java
package br.edu.ifal.leilaogqso;

public class Avaliador {

    private double maiorDeTodos = 0;
    private double menorDeTodos = Double.MAX_VALUE;

    public void avalia(Leilao leilao) {
        for (Lance lance : leilao.getLances()) {
            if (lance.getValor() > maiorDeTodos) {
                maiorDeTodos = lance.getValor();
            }
            
            if (lance.getValor() < menorDeTodos) {
                menorDeTodos = lance.getValor();
            }
        }
    }

    public double getMaiorLance() {
        return maiorDeTodos;
    }

    public double getMenorLance() {
        return menorDeTodos;
    }
}
